Fixed MainController

+++ database/migrations/2014_10_12_000000_create_users_table.php
@@ -20,7 +20,7 @@ return new class extends Migration
             $table->string('last_name');
             $table->string('user_name')->unique();
             $table->string('email')->unique();
-            $table->integer('phone_number');
+            $table->bigInteger('phone_number');
             $table->string('password');
             $table->timestamps();
         });
